feat(smartsheet): implement Phase 2 SmartSheet write operations with bi-directional sync
- Change ProjectEvent.RowNumber from int? to long? to store SmartSheet row IDs
- Add SmartSheetSyncService with sheet creation and row sync operations
- Add SmartSheetSyncController API with sync endpoints and auth validation
- Add sync UI to Project Details page with authentication-aware sync button
- Map events to SmartSheet rows using template column IDs (2455059368464260)
- Support create/update operations based on existing RowNumber values
- Display sync status and refresh timeline after successful sync

+++ src/ShopBoss.Web/Models/ProjectEvent.cs
@@ -23,8 +23,8 @@ public class ProjectEvent
     [MaxLength(100)]
     public string? CreatedBy { get; set; }
     
-    // Optional row number for events imported from SmartSheet
-    public int? RowNumber { get; set; }
+    // SmartSheet row ID for synced events
+    public long? RowNumber { get; set; }
     
     // Optional attachment reference for file/attachment events
     public string? AttachmentId { get; set; }
